misc
More unittests.
Geolocation descriptions are now unique in case they are not some day and MCL
does not like it.

Major change is localization and fix of an error in MCL (obviouly not in this
code):
In "mainlab_chado_loader/includes/class/template/module/mcl_template_phenot"\
+"ype.class.inc", line 224 MCL says they would use the sample_id from the
phenotype-template if provided, but they just don't do it.